pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "todoist-mcp-server"
version = "0.1.0"
description = "A local MCP server for Todoist — task management integration for Claude"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "fastmcp>=2.0.0",
    "todoist-api-python>=2.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
todoist-mcp = "todoist_mcp_server.server:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/todoist_mcp_server"]

[tool.pytest.ini_options]
testpaths = ["tests"]
